Run basic flow test with various options
